Move autostyle button.
