+++ packages/story-editor/src/components/panels/design/textStyle/color.js
@@ -19,34 +19,12 @@
  */
 import PropTypes from 'prop-types';
 import { __ } from '@web-stories-wp/i18n';
-import styled from 'styled-components';
-import { trackEvent } from '@web-stories-wp/tracking';
 
 /**
  * Internal dependencies
  */
-import {
-  Button,
-  BUTTON_SIZES,
-  BUTTON_TYPES,
-  BUTTON_VARIANTS,
-  Icons,
-  themeHelpers,
-} from '@web-stories-wp/design-system';
-import Tooltip from '../../../tooltip';
 import { Color, Row } from '../../../form';
-import useApplyTextAutoStyle from '../../../../utils/useApplyTextAutoStyle';
 import useRichTextFormatting from './useRichTextFormatting';
-
-const AutoStyleButton = styled(Button).attrs({
-  variant: BUTTON_VARIANTS.SQUARE,
-  type: BUTTON_TYPES.TERTIARY,
-  size: BUTTON_SIZES.SMALL,
-})`
-  margin-right: 4px;
-  ${({ theme }) =>
-    themeHelpers.focusableOutlineCSS(theme.colors.border.focus, '#1d1f20')};
-`;
 
 function ColorControls({ selectedElements, pushUpdate, textColorRef }) {
   const {
@@ -54,25 +32,8 @@ function ColorControls({ selectedElements, pushUpdate, textColorRef }) {
     handlers: { handleSetColor },
   } = useRichTextFormatting(selectedElements, pushUpdate);
 
-  const applyTextAutoStyle = useApplyTextAutoStyle(
-    selectedElements[0],
-    (props) => {
-      pushUpdate(props, true);
-      trackEvent('auto_style_text');
-    }
-  );
-
   return (
     <Row>
-      <Tooltip title={__('Adaptive text colors', 'web-stories')}>
-        <AutoStyleButton
-          aria-label={__('Adaptive text colors', 'web-stories')}
-          onClick={applyTextAutoStyle}
-          disabled={selectedElements?.length > 1}
-        >
-          <Icons.ColorBucket />
-        </AutoStyleButton>
-      </Tooltip>
       <Color
         data-testid="text.color"
         value={color}
